@@ -3,21 +3,18 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "axios": "^0.21.1",
-    "bootstrap": "^5.1.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "react-table": "^7.7.0",
+    "@material-ui/core": "^4.12.4",
+    "@material-ui/icons": "^4.11.3",
+    "axios": "^1.1.3",
+    "bootstrap": "^5.2.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.2",
+    "react-scripts": "^5.0.1",
+    "react-table": "^7.8.0",
     "react-table-6": "^6.11.0",
-    "styled-components": "^5.3.0",
-    "web-vitals": "^1.0.1",
-    "@material-ui/core": "^4.11.3",
-    "@material-ui/icons": "^4.11.2"
+    "styled-components": "^5.3.6",
+    "web-vitals": "^3.0.4"
   },
   "scripts": {
     "start": "set PORT=4000 && react-scripts start",
@@ -46,6 +43,5 @@
   "resolutions": {
     "react-scripts/eslint-webpack-plugin": "2.3.0"
   },
-
   "proxy": "http://localhost:3000/api"
 }
